chore: Moves the Menu to views/components folder
