+++ superset-frontend/src/views/components/MenuRight.tsx
